Minor improvements
No more useless traytip, and handles are now closed (not sure if it was ever an issue before, but I guess it's something that should be done anyway). Also removed leading 0s on addresses and removed the weird version offset way of setting mem address values, now just directly using the version value.